Update README.md Hello World Im @Guru22iHub guides pls
i the proper research on cryptocurrencies may require a would-be investor to explore many areas. One area in particular that could prove helpful is simply learning the basic industry terminology. Certain lingo is highly unique to digital currency, making it unlikely that traders would have picked it up when studying other asset classes like stocks, bonds and commodities.

This article will explore the more popular terms and phrases relevant to cryptocurrencies, providing a strong foundation for those interested in exploring this innovative asset class.

Address
In digital currency, an address is basically a destination where a user sends and receives digital currency. In a way, it is similar to a bank account.[1] These addressses usually include a long series of letters and numbers.

Altcoin
An altcoin is a digital currency other than bitcoin. There were more than 1,000 altcoins listed on data source CoinMarketCap at the time of this writing.[2] Another way of describing the term "altcoin" is referring to it as an alternative protocol asset, meaning that it follows a protocol (set of rules) that's different than that of bitcoin.

Arbitrage
In crypto, arbitrage refers to taking advantage of the price difference between two different exchanges. If bitcoin is selling for £8,950 on one exchange and £9,000 on another, a trader can buy the digital currency on the first exchange and sell it on the second for a modest profit.

ATH
"ATH" is an abbreviation of "all-time high." This term can be quite helpful to know for tracking the digital currency markets. These assets are so volatile, so keeping their ATH in mind can prove valuable. A digital currency could potentially hit several local highs before rising to a new all-time high.

Bear/Bearish
"Bears" believe that an asset, for example a digital currency, will decline in value. Another way of putting this is that if a trader thinks a cryptocurrency will depreciate, their sentiment surrounding the digital asset is "bearish." In many situations, traders will make use of this expectation by taking a short position on an asset, meaning that they will make a wager that will pay off should the asset in question fall in value.

Blocks
Many digital currencies make use of blocks, which contain transactions that have been confirmed and then combined together.

Blockchain
The blockchain, which is a distributed ledger system, consists of a series of blocks. These blocks contain verified transactions. The blockchain was designed to be not only decentralised, but also immutable, meaning that entries could not be erased once placed on this distributed ledger. The idea of the blockchain was first introduced when the bitcoin white paper was released in late 2008.[3]

Bull/Bullish
If a trader believes that an asset will rise in value, he or she is a "bull." When an investor has this optimistic expectation of an asset's future bull, this frame of mind is described as "bullish."

Consensus
The network for a digital currency reaches consensus when the network's nodes agree that a transaction took place. This agreement is crucial if the varying network participants (nodes) are to have the same information. In other words, consensus is crucial to distributed ledger systems.

Cryptocurrency
A cryptocurrency is merely a currency that relies on cryptography. Bitcoin, for example, leverages cryptography in order to verify transactions.

Cryptography
Cryptography is basically the process of encoding and decoding information so that would-be observers are unable to understand the information being sent.

DDoS Attack
A distributed denial of service (DDoS) attack takes place when multiple parties work together to overwhelm a system by inundating it with either requests for information or malicious data.[4] Basically, the nefarious parties involved in such an attack want to prevent a resource, such as a server, from being able to provide some specific service, such as serving a web page.